Split provenance enrichment into trustPolicyProvenance.ts
Move the npm-attestations fetch, sigstore.verify call, in-toto statement parsing, dist.integrity comparison, ProvenanceDetails interface, and renderProvenanceDetails out of trustPolicy.ts into a dedicated module. trustPolicy.ts now only knows about ProvenanceDetails as an opaque type and calls fetchAndVerifyProvenance / renderProvenanceDetails through the module boundary.

No behavior change.
