+++ build-tools/packages/build-cli/src/commands/check/trustPolicy.ts
@@ -4,7 +4,6 @@
  */
 
 import { mkdir, rm, writeFile } from "node:fs/promises";
-import { createRequire } from "node:module";
 import path from "node:path";
 
 import { Flags } from "@oclif/core";
@@ -14,34 +13,11 @@ import pacote from "pacote";
 import semver from "semver";
 
 import { BaseCommandWithBuildProject } from "../../library/commands/base.js";
-
-/**
- * Minimal runtime interface for the `sigstore` package's verify API.
- *
- * `sigstore` is loaded via `createRequire` rather than a static `import` so
- * TypeScript does not pull its transitive `.d.ts` graph into our program.
- * That graph (`@sigstore/core`, `@sigstore/sign`, `@sigstore/tuf`,
- * `make-fetch-happen`, `@sigstore/rekor-types`) has type errors against this
- * repo's `@types/node` baseline that we cannot fix from here.
- */
-interface SigstoreModule {
-	verify(bundle: SigstoreBundle): Promise<SigstoreSigner>;
-}
-
-/** Subset of `@sigstore/bundle`'s `SerializedBundle` we actually read. */
-interface SigstoreBundle {
-	dsseEnvelope?: { payload?: string };
-}
-
-/** Subset of `@sigstore/verify`'s `Signer` we actually read. */
-interface SigstoreSigner {
-	identity?: {
-		subjectAlternativeName?: string;
-		extensions?: { issuer?: string };
-	};
-}
-
-const sigstore = createRequire(import.meta.url)("sigstore") as SigstoreModule;
+import {
+	fetchAndVerifyProvenance,
+	renderProvenanceDetails,
+	type ProvenanceDetails,
+} from "./trustPolicyProvenance.js";
 
 /**
  * The error code pnpm emits (both as the top-level `code` and inside `err`)
@@ -66,46 +42,6 @@ interface PinnedVersion {
  * `none` is our own sentinel for "neither."
  */
 type TrustEvidence = "none" | "provenance" | "trustedPublisher";
-
-/**
- * Verified provenance metadata extracted from a Sigstore SLSA bundle
- * served by the npm registry's attestations endpoint.
- *
- * All fields are populated only after the bundle has been cryptographically
- * verified via `sigstore.verify` (which checks the Fulcio-issued certificate
- * chain, the Rekor inclusion proof, and the DSSE signature). A missing field
- * means the bundle did not carry that piece of metadata, not that it failed
- * verification — verification failures land on `verificationError` instead
- * and leave the rest of the fields undefined.
- */
-interface ProvenanceDetails {
-	/** Signing identity from the Fulcio certificate's SAN + issuer extension. */
-	signerIdentity?: { issuer?: string; subject?: string };
-	/** Source repository URL (e.g. `https://github.com/owner/repo`). */
-	sourceRepo?: string;
-	/** Git ref the build ran on (e.g. `refs/tags/v1.2.3`). */
-	sourceRef?: string;
-	/** Git commit SHA the build ran on. */
-	commit?: string;
-	/** Path of the build workflow file (e.g. `.github/workflows/release.yml`). */
-	workflowPath?: string;
-	/** SLSA builder ID (e.g. `https://github.com/actions/runner/github-hosted`). */
-	builderId?: string;
-	/** URL of the build run that produced the artifact. */
-	runUrl?: string;
-	/**
-	 * `true` if the SLSA statement's subject digest matches the registry
-	 * manifest's `dist.integrity`. `false` if they disagree (a strong signal
-	 * the registry tarball does not match what was attested). `undefined` if
-	 * the comparison could not be performed (algorithm mismatch, etc.).
-	 */
-	subjectDigestVerified?: boolean;
-	/**
-	 * Populated when fetching or verifying the bundle failed. The other
-	 * fields are left undefined in that case.
-	 */
-	verificationError?: string;
-}
 
 /**
  * A trust-downgrade event extracted from pnpm's NDJSON stream.
@@ -879,12 +815,17 @@ async function enrichViolations(
 			// `provenanceDetails.verificationError` and reported alongside
 			// the rest of the violation.
 			if (v.evidence === "provenance") {
-				v.provenanceDetails = await enrichProvenance(v.name, v.version, offender, verbose);
+				v.provenanceDetails = await fetchAndVerifyProvenance(
+					v.name,
+					v.version,
+					offender,
+					verbose,
+				);
 			}
 			if (v.priorTrusted?.evidence === "provenance") {
 				const ptManifest = versions[v.priorTrusted.version];
 				if (ptManifest !== undefined) {
-					v.priorTrusted.provenanceDetails = await enrichProvenance(
+					v.priorTrusted.provenanceDetails = await fetchAndVerifyProvenance(
 						v.name,
 						v.priorTrusted.version,
 						ptManifest,
@@ -896,163 +837,6 @@ async function enrichViolations(
 			v.enrichmentError = err instanceof Error ? err.message : String(err);
 		}
 	}
-}
-
-/**
- * Subset of the npm registry's attestation envelope. Each entry carries a
- * `predicateType` identifying the in-toto statement kind, plus the raw
- * Sigstore bundle. We care about the SLSA provenance entry; the npm
- * publish-attestation entry (predicateType
- * `https://github.com/npm/attestation/...`) is ignored.
- */
-interface RegistryAttestation {
-	predicateType: string;
-	bundle: SigstoreBundle;
-}
-
-interface RegistryAttestationsResponse {
-	attestations?: RegistryAttestation[];
-}
-
-/** SLSA v1 in-toto statement subset. Other predicate types share the envelope. */
-interface InTotoStatement {
-	subject?: { name?: string; digest?: Record<string, string> }[];
-	predicateType?: string;
-	predicate?: {
-		buildDefinition?: {
-			externalParameters?: {
-				workflow?: { ref?: string; repository?: string; path?: string };
-			};
-			resolvedDependencies?: {
-				uri?: string;
-				digest?: Record<string, string>;
-			}[];
-		};
-		runDetails?: {
-			builder?: { id?: string };
-			metadata?: { invocationId?: string };
-		};
-	};
-}
-
-/**
- * Fetches the SLSA provenance attestation for `name@version` from the npm
- * registry, verifies the Sigstore bundle, and extracts the source-repo,
- * workflow, and signer-identity fields useful for auditing a downgrade.
- *
- * Always returns a `ProvenanceDetails`. On failure it returns one with only
- * `verificationError` set; the caller renders that alongside the other
- * violation fields so a single bad attestation can't abort the audit.
- *
- * Network access is hardcoded to `https://registry.npmjs.org`. The
- * attestations endpoint is npm-public-only; private/scoped registries
- * (Artifactory, ADO, GitHub Packages) do not implement it. If the lookup
- * 404s we report that explicitly so the user knows it's a registry-API
- * limitation, not a verification failure.
- */
-async function enrichProvenance(
-	name: string,
-	version: string,
-	manifest: pacote.Manifest & TrustFields,
-	verbose: (msg: string) => void,
-): Promise<ProvenanceDetails> {
-	try {
-		verbose(`    fetching attestations for ${name}@${version}`);
-		const url = `https://registry.npmjs.org/-/npm/v1/attestations/${encodeURIComponent(name)}@${encodeURIComponent(version)}`;
-		const res = await fetch(url);
-		if (res.status === 404) {
-			return { verificationError: "no attestations served (registry returned 404)" };
-		}
-		if (!res.ok) {
-			return { verificationError: `HTTP ${res.status} fetching ${url}` };
-		}
-		const body = (await res.json()) as RegistryAttestationsResponse;
-		const slsa = body.attestations?.find((a) =>
-			a.predicateType?.startsWith("https://slsa.dev/provenance/"),
-		);
-		if (slsa === undefined) {
-			return { verificationError: "no SLSA provenance entry in registry attestations" };
-		}
-		// `verify` (no payload overload) checks the Fulcio cert chain, the
-		// Rekor inclusion proof, and the DSSE signature against the embedded
-		// statement. It does not need the tarball because the statement is
-		// inside the DSSE envelope.
-		const signer = await sigstore.verify(slsa.bundle);
-		const statement = parseInTotoStatement(slsa.bundle);
-		if (statement === undefined) {
-			return { verificationError: "verified bundle but could not parse in-toto statement" };
-		}
-		const wf = statement.predicate?.buildDefinition?.externalParameters?.workflow;
-		const builder = statement.predicate?.runDetails?.builder?.id;
-		const runUrl = statement.predicate?.runDetails?.metadata?.invocationId;
-		// Find the resolved-dependency entry that points at the source repo;
-		// SLSA v1 GitHub builders put the source there with a `gitCommit` digest.
-		const sourceDep = statement.predicate?.buildDefinition?.resolvedDependencies?.find(
-			(d) => d.digest?.gitCommit !== undefined,
-		);
-		return {
-			signerIdentity: {
-				issuer: signer.identity?.extensions?.issuer,
-				subject: signer.identity?.subjectAlternativeName,
-			},
-			sourceRepo: wf?.repository,
-			sourceRef: wf?.ref,
-			commit: sourceDep?.digest?.gitCommit,
-			workflowPath: wf?.path,
-			builderId: builder,
-			runUrl,
-			subjectDigestVerified: compareSubjectDigest(manifest, statement),
-		};
-	} catch (err) {
-		return { verificationError: err instanceof Error ? err.message : String(err) };
-	}
-}
-
-/**
- * Decodes the DSSE envelope payload from a Sigstore bundle as a UTF-8 JSON
- * in-toto statement. Returns `undefined` if the bundle has no DSSE envelope
- * (e.g. a `messageSignature`-style bundle, which is not used for npm
- * provenance) or if the payload doesn't parse as JSON.
- */
-function parseInTotoStatement(bundle: SigstoreBundle): InTotoStatement | undefined {
-	const env = bundle.dsseEnvelope;
-	if (env?.payload === undefined) return undefined;
-	try {
-		return JSON.parse(Buffer.from(env.payload, "base64").toString("utf8")) as InTotoStatement;
-	} catch {
-		return undefined;
-	}
-}
-
-/**
- * Compares the in-toto statement's first subject digest against the
- * registry manifest's `dist.integrity`. Returns `true` if they agree
- * (the tarball matches what was attested), `false` if they disagree
- * (a strong red flag), and `undefined` if the comparison can't be made
- * because the algorithms differ or either side is missing.
- *
- * `dist.integrity` is `<algo>-<base64>`; the SLSA digest is hex.
- */
-function compareSubjectDigest(
-	manifest: { dist?: { integrity?: string } },
-	statement: InTotoStatement,
-): boolean | undefined {
-	const integrity = manifest.dist?.integrity;
-	const subjectDigest = statement.subject?.[0]?.digest;
-	if (integrity === undefined || subjectDigest === undefined) return undefined;
-	const dashIdx = integrity.indexOf("-");
-	if (dashIdx <= 0) return undefined;
-	const algo = integrity.slice(0, dashIdx);
-	const expectedBase64 = integrity.slice(dashIdx + 1);
-	const actualHex = subjectDigest[algo];
-	if (actualHex === undefined) return undefined;
-	let actualBase64: string;
-	try {
-		actualBase64 = Buffer.from(actualHex, "hex").toString("base64");
-	} catch {
-		return undefined;
-	}
-	return actualBase64 === expectedBase64;
 }
 
 /** Formats an ISO timestamp as `YYYY-MM-DD`, or undefined if invalid/missing. */
@@ -1073,49 +857,5 @@ function formatEvidence(evidence: TrustEvidence | undefined): string | undefined
 			return "no trust evidence";
 		default:
 			return undefined;
-	}
-}
-
-/**
- * Emits a human-readable rendering of {@link ProvenanceDetails} as one
- * `key: value` line per populated field, prefixed by `indent`. Skipped
- * fields produce no output. On verification failure only the error line
- * is printed so the report stays compact.
- */
-function renderProvenanceDetails(
-	indent: string,
-	details: ProvenanceDetails,
-	emit: (line: string) => void,
-): void {
-	if (details.verificationError !== undefined) {
-		emit(`${indent}provenance: ${details.verificationError}`);
-		return;
-	}
-	emit(`${indent}provenance (verified):`);
-	if (details.sourceRepo !== undefined) {
-		const refSuffix = details.sourceRef === undefined ? "" : ` @ ${details.sourceRef}`;
-		emit(`${indent}  source: ${details.sourceRepo}${refSuffix}`);
-	}
-	if (details.commit !== undefined) {
-		emit(`${indent}  commit: ${details.commit}`);
-	}
-	if (details.workflowPath !== undefined) {
-		emit(`${indent}  workflow: ${details.workflowPath}`);
-	}
-	if (details.builderId !== undefined) {
-		emit(`${indent}  builder: ${details.builderId}`);
-	}
-	if (details.runUrl !== undefined) {
-		emit(`${indent}  run: ${details.runUrl}`);
-	}
-	if (details.signerIdentity?.subject !== undefined) {
-		const issuerSuffix =
-			details.signerIdentity.issuer === undefined
-				? ""
-				: ` (issuer: ${details.signerIdentity.issuer})`;
-		emit(`${indent}  signer: ${details.signerIdentity.subject}${issuerSuffix}`);
-	}
-	if (details.subjectDigestVerified === false) {
-		emit(`${indent}  WARNING: attested subject digest does not match registry tarball`);
 	}
 }
